Made balanceAPI component; Chart can show
